+++ application/src/main/java/it/smartcommunitylabdhub/core/services/DataItemServiceImpl.java
@@ -404,10 +404,29 @@ public class DataItemServiceImpl
     }
 
     @Override
-    public void deleteDataItem(@NotNull String id) {
+    public void deleteDataItem(@NotNull String id, @Nullable Boolean cascade) {
         log.debug("delete dataItem with id {}", String.valueOf(id));
         try {
-            entityService.delete(id);
+            DataItem dataItem = entityService.find(id);
+            if (dataItem != null) {
+                if (Boolean.TRUE.equals(cascade)) {
+                    //files
+                    log.debug("cascade delete files for dataItem with id {}", String.valueOf(id));
+
+                    //extract path from spec
+                    DataItemBaseSpec spec = new DataItemBaseSpec();
+                    spec.configure(dataItem.getSpec());
+
+                    String path = spec.getPath();
+                    if (StringUtils.hasText(path)) {
+                        //delete files
+                        filesService.remove(path, UserAuthenticationHelper.getUserAuthentication());
+                    }
+                }
+
+                //delete entity
+                entityService.delete(id);
+            }
         } catch (StoreException e) {
             log.error("store error: {}", e.getMessage());
             throw new SystemException(e.getMessage());
@@ -415,27 +434,39 @@ public class DataItemServiceImpl
     }
 
     @Override
-    public void deleteDataItems(@NotNull String project, @NotNull String name) {
+    public void deleteDataItems(@NotNull String project, @NotNull String name, @Nullable Boolean cascade) {
         log.debug("delete dataItems for project {} with name {}", project, name);
-
-        Specification<DataItemEntity> spec = Specification.allOf(
-            CommonSpecification.projectEquals(project),
-            CommonSpecification.nameEquals(name)
-        );
-        try {
-            long count = entityService.deleteAll(spec);
-            log.debug("deleted count {}", count);
-        } catch (StoreException e) {
-            log.error("store error: {}", e.getMessage());
-            throw new SystemException(e.getMessage());
+        if (Boolean.TRUE.equals(cascade)) {
+            //delete one by one with cascade
+            findDataItems(project, name).forEach(d -> deleteDataItem(d.getId(), Boolean.TRUE));
+        } else {
+            //bulk delete entities only
+            Specification<DataItemEntity> spec = Specification.allOf(
+                CommonSpecification.projectEquals(project),
+                CommonSpecification.nameEquals(name)
+            );
+            try {
+                long count = entityService.deleteAll(spec);
+                log.debug("bulk deleted count {}", count);
+            } catch (StoreException e) {
+                log.error("store error: {}", e.getMessage());
+                throw new SystemException(e.getMessage());
+            }
         }
     }
 
     @Override
-    public void deleteDataItemsByProject(@NotNull String project) {
+    public void deleteDataItemsByProject(@NotNull String project, @Nullable Boolean cascade) {
         log.debug("delete dataItems for project {}", project);
         try {
-            entityService.deleteAll(CommonSpecification.projectEquals(project));
+            if (Boolean.TRUE.equals(cascade)) {
+                //delete one by one with cascade
+                entityService
+                    .searchAll(CommonSpecification.projectEquals(project))
+                    .forEach(d -> deleteDataItem(d.getId(), Boolean.TRUE));
+            } else {
+                entityService.deleteAll(CommonSpecification.projectEquals(project));
+            }
         } catch (StoreException e) {
             log.error("store error: {}", e.getMessage());
             throw new SystemException(e.getMessage());
